Rename pulsarAdminConsole and pulsarHeartbeat (#2)
* pulsarAdminConsole

* pulsar-heartbeat

* update kubernetes version for kind testing

+++ tests/e2e-kind.sh
@@ -7,7 +7,7 @@ set -o pipefail
 readonly CT_VERSION=v3.3.1
 readonly KIND_VERSION=v0.9.0
 readonly CLUSTER_NAME=pulsar-helm-test
-readonly K8S_VERSION=v1.16.15
+readonly K8S_VERSION=v1.18.8
 
 run_ct_container() {
     echo 'Running ct container...'
@@ -32,10 +32,12 @@ docker_exec() {
 }
 
 create_kind_cluster() {
-    echo 'Installing kind...'
-    curl -Lo ./kind https://kind.sigs.k8s.io/dl/$KIND_VERSION/kind-linux-amd64
-    chmod +x ./kind
-    sudo mv kind /usr/local/bin/kind
+    if [ ! -f /usr/local/bin/kind ]; then
+        echo 'Installing kind...'
+        curl -Lo ./kind https://kind.sigs.k8s.io/dl/$KIND_VERSION/kind-linux-amd64
+        chmod +x ./kind
+        sudo mv kind /usr/local/bin/kind
+    fi
 
     kind create cluster --name "$CLUSTER_NAME" --config tests/kind-config.yaml --image "kindest/node:$K8S_VERSION" --wait 60s
